# child_process: Fix regression on Windows for FAT filesystems