This fixes a regression caused by https://github.com/ziglang/zig/pull/13993

As an optimization, the first call to `NtQueryDirectoryFile` would only ask for a single result and assume that if the result returned did not match the app_name exactly, then the unappended app_name did not exist. However, this relied on the assumption that the unappended app_name would always be returned first, but that only seems to be the case on NTFS. On FAT filesystems, the order of returned files can be different, which meant that it could assume the unappended file doesn't exist when it actually does.

This commit fixes that by fully iterating the wildcard matches via `NtQueryDirectoryFile` and taking note of any unappended/PATHEXT-appended filenames it finds. In practice, this strategy does not introduce a speed regression compared to the previous (buggy) implementation.

```
Benchmark 1 (10 runs): winpathbench-master.exe
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           508ms ± 4.08ms     502ms …  517ms          1 (10%)        0%
  peak_rss           3.62MB ± 2.76KB    3.62MB … 3.63MB          0 ( 0%)        0%
Benchmark 2 (10 runs): winpathbench-fat32-fix.exe
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           500ms ± 21.4ms     480ms …  535ms          0 ( 0%)          -  1.5% ±  2.8%
  peak_rss           3.62MB ± 2.76KB    3.62MB … 3.63MB          0 ( 0%)          -  0.0% ±  0.1%
```
(same benchmark as https://github.com/ziglang/zig/pull/13993, but different `PATH` environment variable)

---

Partially addresses #16374 (it fixes `zig build` on FAT32 when no `zig-cache` is present, but `error.Unexpected NTSTATUS=0xc0000035` still happens during `zig build` once `zig-cache` is created)

Before (where `E:\` is a FAT32 filesystem):

```
E:\zigtest> zig build
error: FileNotFound
```

After this commit:

```
E:\zigtest> zig build

E:\zigtest> ".\zig-out\bin\zigtest.exe"
All your codebase are belong to us.
Run `zig build test` to run the tests.
```

---

Note: there's a fairly comprehensive standalone test for Windows spawn behavior, so as long as the tests pass we can have some confidence that this won't regress the spawn behavior:

https://github.com/ziglang/zig/blob/master/test/standalone/windows_spawn/main.zig